+++ crypto-onramp/src/main/java/com/stripe/android/crypto/onramp/analytics/OnrampAnalyticsEvent.kt
@@ -102,6 +102,14 @@ internal sealed class OnrampAnalyticsEvent(
         name = "terms_and_conditions_completed"
     )
 
+    data object TermsOfServiceStarted : OnrampAnalyticsEvent(
+        name = "terms_of_service_started"
+    )
+
+    data object TermsOfServiceCompleted : OnrampAnalyticsEvent(
+        name = "terms_of_service_completed"
+    )
+
     class WalletRegistered(
         network: CryptoNetwork
     ) : OnrampAnalyticsEvent(
@@ -260,7 +268,8 @@ internal sealed class OnrampAnalyticsEvent(
             RetrieveMissingIdentifiers("retrieve_missing_identifiers"),
             SubmitIdentifiers("submit_identifiers"),
             PresentUserAttestation("present_user_attestation"),
-            PresentTermsAndConditionsIfNeeded("present_terms_and_conditions_if_needed")
+            PresentTermsAndConditionsIfNeeded("present_terms_and_conditions_if_needed"),
+            PresentTermsOfServiceIfNeeded("present_terms_of_service_if_needed"),
         }
     }